apps/api/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Celery
celery[redis]==5.3.6
redis==5.0.1

# Audio Processing
pydub==0.25.1
numpy==1.26.3
soundfile==0.12.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
torchaudio==2.1.2+cpu

# OpenAI (for Whisper API)
openai>=1.12.0

# AWS S3
boto3==1.34.14

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
